stats calculated and shared with player, healt regen

+++ gameserver/src/main/java/com/shnok/javaserver/model/template/EntityTemplate.java
@@ -5,12 +5,12 @@ import lombok.Data;
 
 @Data
 public class EntityTemplate {
-    public byte baseSTR;
-    public byte baseCON;
-    public byte baseDEX;
-    public byte baseINT;
-    public byte baseWIT;
-    public byte baseMEN;
+    public int baseSTR;
+    public int baseCON;
+    public int baseDEX;
+    public int baseINT;
+    public int baseWIT;
+    public int baseMEN;
     public int baseHpMax;
     public int baseCpMax;
     public int baseMpMax;
